book->getbook으로 기능이전 + 썸네일표시
