+++ kh7semimain/src/library/servlet/recommend/BookRecommendInsertServlet.java
@@ -1,4 +1,4 @@
-package library.servlet.book;
+package library.servlet.recommend;
 
 import java.io.IOException;
 
@@ -9,29 +9,26 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import library.beans.RecommendDao;
-import library.beans.WishlistDao;
-import library.beans.WishlistDto;
+import library.beans.RecommendDto;
 
-@WebServlet(urlPatterns ="/book/bookWishlistInsert.kh")
-public class BookWishlistInsertServlet extends HttpServlet{
+@WebServlet(urlPatterns ="/recommend/recommendInsert.kh")
+public class BookRecommendInsertServlet extends HttpServlet{
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		try {
 			//준비 : clientNo, bookIsbn
 			req.setCharacterEncoding("UTF-8");
-			WishlistDto wishlistDto = new WishlistDto();
-			wishlistDto.setClientNo(Integer.parseInt(req.getParameter("clientNo"))); 
-			wishlistDto.setBookIsbn(req.getParameter("bookIsbn"));
+			RecommendDto recommendDto = new RecommendDto();
+			recommendDto.setClientNo(Integer.parseInt(req.getParameter("clientNo"))); 
+			recommendDto.setBookIsbn(req.getParameter("bookIsbn"));
 			
 			//처리
-			WishlistDao wishlistDao = new WishlistDao();
-			int wishlistNo = wishlistDao.getSequence();//게시글번호(DB시퀀스)
-			wishlistDto.setWishlistNo(wishlistNo);
-			
-			wishlistDao.insert(wishlistDto);
+			RecommendDao recommendDao = new RecommendDao();
+			recommendDao.insert(recommendDto);
+			String root = req.getContextPath();
 			
 			//출력 : 책 리스트로 복귀
-			resp.sendRedirect("bookList.jsp");
+			resp.sendRedirect(root+"/search/searchList.jsp");
 			
 		}
 		catch(Exception e) {
