+++ kh7semimain/src/library/servlet/wishlist/wishlistInsertServlet.java
@@ -1,4 +1,4 @@
-package library.servlet.book;
+package library.servlet.wishlist;
 
 import java.io.IOException;
 
@@ -8,14 +8,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-
 import library.beans.WishlistDao;
 import library.beans.WishlistDto;
 
-@WebServlet(urlPatterns= {
-		"/book/bookWishlistDelete.kh"
-})
-public class BookWishlistDeleteServlet extends HttpServlet{
+@WebServlet(urlPatterns="/wishlist/wishlistInsert.kh")
+public class wishlistInsertServlet extends HttpServlet{
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		try {
@@ -27,13 +24,14 @@ public class BookWishlistDeleteServlet extends HttpServlet{
 			
 			//처리
 			WishlistDao wishlistDao = new WishlistDao();
-			wishlistDao.delete(wishlistDto.getClientNo(), wishlistDto.getBookIsbn());
+			int wishlistNo = wishlistDao.getSequence();//게시글번호(DB시퀀스)
+			wishlistDto.setWishlistNo(wishlistNo);
 			
+			wishlistDao.insert(wishlistDto);
+			String root = req.getContextPath();
 			
-			//출력 : 책 리스트로 복귀
-		
-			resp.sendRedirect("bookList.jsp");	
-		
+			resp.sendRedirect(root+"/search/searchList.jsp");
+			
 		}
 		catch(Exception e) {
 			e.printStackTrace();
